Add simple route learning on a per hop basis
This will not discover MAC addresses on I2C and CAN, but works well over point-to-point links
in promiscous mode, like KISS and Spacelink. This will break communication as we know it
since subsystems which receive a packet from a ground-station node over I2C will insert
a direct route to the ground-station node using the I2C interface, but without a valid MAC
nexthop address. Maybe this could be solved using some kind of ARP on I2C, so that any
node that knows a route to the address requested on I2C, will reply with its MAC address.
On CAN it's a bit different, since the routers are in promiscous mode and automatically
will pick up traffic on the network and route it to the ground.